fix out-of-bounds when writing bitmask at the bottom-right corner of the canvas
When writing a span of bits whose right edge is aligned with the boundary
between two u64s, Rasterizer::fill_cells would perform a no-op write of all
zeroes into the u64 after the end of the span. This was harmless except in the
case where the right edge of the span was aligned with the bottom-right corner
of the canvas, in which case that no-op write would result in an out-of-bounds
access.

Now, in the case described above (when the right edge of a span of bits is
aligned with the boundary between u64s), the unconditional "remainder" write at
the end of fill_cells writes to the final non-empty u64 rather than performing
a no-op write to the next u64. This fixes the out-of-bounds access when the
right edge is aligned with the bottom-right corner of the canvas.